requirements.txt
pillow
django-cleanup
django-allauth
django-htmx
